+++ src/config/laravel-backup.php
@@ -37,7 +37,7 @@ return [
              * Currently only MySQL- and PostgreSQL-databases are supported.
              */
             'databases' => [
-                'mysql'
+                'mysql',
             ],
         ],
 
@@ -90,10 +90,9 @@ return [
              * After cleaning up the backups remove the oldest backup until
              * this amount of megabytes has been reached.
              */
-            'deleteOldestBackupsWhenUsingMoreMegabytesThan' => 5000
-        ]
+            'deleteOldestBackupsWhenUsingMoreMegabytesThan' => 5000,
+        ],
     ],
-
 
     /*
      *  In this array you can specify which backups should be monitored.
@@ -102,9 +101,9 @@ return [
      */
     'monitorBackups' => [
         [
-            'name' => env('APP_URL'),
-            'disks' => ['backups'],
-            'newestBackupsShouldNotBeOlderThanDays' => 1,
+            'name'                                   => env('APP_URL'),
+            'disks'                                  => ['backups'],
+            'newestBackupsShouldNotBeOlderThanDays'  => 1,
             'storageUsedMayNotBeHigherThanMegabytes' => 5000,
         ],
 
@@ -137,7 +136,7 @@ return [
             'whenHealthyBackupWasFound'   => ['log'],
             'whenBackupHasFailed'         => ['log', 'mail'],
             'whenCleanupHasFailed'        => ['log', 'mail'],
-            'whenUnHealthyBackupWasFound' => ['log', 'mail']
+            'whenUnHealthyBackupWasFound' => ['log', 'mail'],
         ],
 
         /*
@@ -156,5 +155,5 @@ return [
             'username' => 'Backup bot',
             'icon'     => ':robot:',
         ],
-    ]
+    ],
 ];